Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/dom/dynamic-markup-insertion/opening-the-input-stream/document.open-03-expected.txt


FAIL document.open and singleton replacement assert_equals: setTimeout expected false but got true
